chore: extract image build to the image package
